backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
langchain
langchain-community
langchain-core
langchain-google-genai
google-generativeai>=0.5.0
langchain-chroma
langchain-huggingface
sentence-transformers
chromadb
langchain-groq
faiss-cpu
pypdf